refactor(web): split api/client.ts into domain modules
The 1238-line src/api/client.ts is replaced by a client/ package with one
module per endpoint group (team, observability, misc, wiki, agents,
scheduler, mcp, settings) plus _shared.ts for the common ApiValidationError /
parseDetailOrThrow helpers. client/index.ts re-exports everything, so all 69
consumers keep importing named symbols from '@/api/client' unchanged.

Pure move — no behavior change. Verified: tsc -b, tsc -p tsconfig.test.json,
eslint, and the full 2156-test web suite all pass.
